feat: separate months and languages
